backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-mpd2==3.1.1
Pillow==10.4.0
requests==2.32.3
gunicorn==23.0.0
